+++ src/main/java/com/bew/demo/dao/FileImageRepository.java
@@ -11,7 +11,7 @@ import com.bew.demo.model.FileImage;
 
 
 @Repository
-public interface FileImageRepository extends JpaRepository <FileImage, Integer  >{
+public interface FileImageRepository extends JpaRepository <FileImage, Long>{
 
 	  
     @Query(
