+++ src/main/java/com/bew/demo/restController/DocsDictamenRestController.java
@@ -39,7 +39,7 @@ public class DocsDictamenRestController {
 
     @GetMapping("/getFile/{idFile}")
     @ResponseBody
-    public ResponseEntity<ByteArrayResource> serveFile(@PathVariable Integer idFile) throws EmptyResultException {
+    public ResponseEntity<ByteArrayResource> serveFile(@PathVariable Long idFile) throws EmptyResultException {
    	
         return  docsDictamenService.load(idFile);
     }
