+++ src/main/java/com/bew/demo/service/DictamenService.java
@@ -8,10 +8,10 @@ import com.bew.demo.exception.EmptyResultException;
 public interface DictamenService {
 
 	List<DictamenDTO> findAll();
-	DictamenDTO findById(Integer idDictamen);
+	DictamenDTO findById(Long idDictamen);
 	void saveDictamen(DictamenDTO dictamenDTO);
 	void updateDictamen(DictamenDTO dictamenDTO)throws EmptyResultException;
-	void deleteDictamen(Integer idDictamen)throws EmptyResultException;
+	void deleteDictamen(Long idDictamen)throws EmptyResultException;
 	DictamenDTO findByIdAlumno(Long idAlumno);
 	List<DictamenDTO> findByEstado(String estado);
