+++ src/main/java/com/bew/demo/service/DocsBajaService.java
@@ -8,7 +8,7 @@ import com.bew.demo.exception.EmptyResultException;
 
 public interface DocsBajaService {
 	
-	ResponseEntity<ByteArrayResource> load(Integer fileId) throws EmptyResultException;
+	ResponseEntity<ByteArrayResource> load(Long fileId) throws EmptyResultException;
 
 	String store(MultipartFile file, String idDoc) throws EmptyResultException;
