+++ src/main/java/com/bew/demo/service/DocsDictamenServiceImpl.java
@@ -37,7 +37,7 @@ public class DocsDictamenServiceImpl implements DocsDictamenService {
 }
 
 	@Override
-	public ResponseEntity<ByteArrayResource> load(Integer idFile) throws EmptyResultException {
+	public ResponseEntity<ByteArrayResource> load(Long idFile) throws EmptyResultException {
 		DocsDictamen file = docsDictamenRepository.findById(idFile).orElseThrow(() -> new   EmptyResultException("File not found with id " + idFile));
