+++ src/main/java/com/bew/demo/service/DocsServicioServiceImpl.java
@@ -37,7 +37,7 @@ public class DocsServicioServiceImpl implements DocsServicioService{
 }
 
 	@Override
-	public ResponseEntity<ByteArrayResource> load(Integer idFile) throws EmptyResultException {
+	public ResponseEntity<ByteArrayResource> load(Long idFile) throws EmptyResultException {
 		DocsServicio file = docsServicioRepository.findById(idFile).orElseThrow(() -> new   EmptyResultException("File not found with id " + idFile));
